# target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: API de Votação
  version: 1.0.0
  description: API para gerenciar votações de pautas.
paths:
...
  /pautas/{pautaId}/votos:
    post:
      summary: Votar
      description: Registra um voto em uma pauta específica.
      parameters:
        - name: pautaId
          in: path
          required: true
          description: ID da pauta na qual o voto será registrado
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Voto'
      responses:
        '200':
          description: Voto registrado com sucesso.
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
...
    Voto:
      type: object
      properties:
        usuarioId:
          type: integer
        escolha:
          type: string
          enum: [SIM, NAO]